pyproject.toml
[tool.poetry]
name = "feel"
version = "0.1.0"
description = "This a simple social media app I'm making to learn web development"
authors = ["Mati <mathiaskassaye8@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
django = "==3.1"
graphene-django = "==2.12.1"
psycopg2 = "==2.8.5"
django-graphql-jwt = "==0.3.1"
graphene-django-extras = "==0.4.9"
aniso8601 = "==7.0.0"
asgiref = "==3.2.10"
django-filter = "==2.3.0"
graphene = "==2.1.8"
graphql-core = "==2.3.2"
graphql-relay = "==2.0.1"
promise = "==2.3"
python-dateutil = "==2.8.1"
pytz = "==2020.1"
singledispatch = "==3.4.0.3"
six = "==1.15.0"
sqlparse = "==0.3.1"
unidecode = "==1.1.1"
PyJWT = "==1.7.1"
Rx = "==1.6.1"
dj-database-url = "*"
gunicorn = "*"
django-heroku = "*"
whitenoise = {extras = ["brotli"], version = "*"}
django-notifications-hq = "*"
graphene-file-upload = "*"
django-cloudinary-storage = "*"
cloudinary = "*"
graphene-subscriptions = "*"
channels = "*"
pillow = "*"
pyOpenSSL = "^19.1.0"
wagtail = "^2.10.2"
service_identity = "^18.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
